requirements.txt
pandas==1.5.3
Pillow==10.2.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
transformers==4.37.1
